+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Changed the role handling from masterdata to vbase to avoid duplicated entries
+
 ## [1.13.0] - 2022-02-02
 
 ### Added